\name{rivers}
\alias{rivers}
\docType{data}
\title{River lines (Norway)}
\description{River lines (Norway)}
\usage{data("rivers")}
\format{
  The format is:
Formal class 'SpatialLinesDataFrame' [package "sp"] with 4 slots
  ..@ data       :'data.frame':	97 obs. of  11 variables:
  .. ..$ SL_ID     : chr [1:97] "0" "1" "2" "3" ...
  .. ..$ FNODE_    : num [1:97] 3 5 7 9 1 14 14 15 17 18 ...
  .. ..$ TNODE_    : num [1:97] 2 4 8 10 13 13 12 16 11 14 ...
  .. ..$ LPOLY_    : num [1:97] 2 2 1 2 2 2 2 2 2 2 ...
  .. ..$ RPOLY_    : num [1:97] 2 1 2 2 2 2 2 2 2 2 ...
  .. ..$ LENGTH    : num [1:97] 1429 7370 11041 8412 153450 ...
  .. ..$ VALINJE_  : num [1:97] 1 2 3 4 5 6 7 8 9 10 ...
  .. ..$ VALINJE_ID: num [1:97] 97 97 64 71 135 66 133 73 69 66 ...
  .. ..$ LTEMA     : num [1:97] 3211 3211 3211 3211 3211 ...
  .. ..$ VANNBR    : num [1:97] 13 13 13 13 13 13 13 13 13 13 ...
  .. ..$ DATO      : num [1:97] 2e+07 2e+07 2e+07 2e+07 2e+07 ...
  ..@ lines      :List of 97
  ..@ bbox       : num [1:2, 1:2] 5.84 58.02 30.15 70.2
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:2] "x" "y"
  .. .. ..$ : chr [1:2] "min" "max"
  ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. ..@ projargs: chr "+proj=longlat +ellps=WGS84 +no_defs"
}
\examples{
data(rivers)
}
\keyword{datasets}
